corrects dependancy mapping for all commands
