+++ public/app/commands/bin/clear/clear.js
@@ -2,8 +2,8 @@ var React = require('react')
 var $ = window.jQuery = window.$ = require('jquery')
 var Ascii = require('../../../components/ascii')
 var CommandActions = require('../../../actions/command-actions')
-var expectOption = require('../../../lib/command-parser').expectOption
-var expectOptionWithArgument = require('../../../lib/command-parser').expectOptionWithArgument
+var expectOption = require('../../../lib/command-helper').expectOption
+var expectOptionWithArgument = require('../../../lib/command-helper').expectOptionWithArgument
 
 exports.description = 'Clears your console window'
 exports.name = 'clear'
@@ -14,6 +14,6 @@ exports.execute = function(args) {
 exports.help = function() {
   return [
     <Ascii key={0} value='clear' font='bell'/>,
-    <Markdown file='/markdown/commands/manifesto.md'/>
+    <Markdown file={'/app/commands/bin/' + this.name + '/help.md'}/>
   ]
 }
