+++ public/app/commands/bin/clear/help.md
@@ -1,0 +1,1 @@
+## CRAZY HELP
