+++ public/app/commands/bin/ls/ls.js
@@ -1,12 +1,14 @@
 var React = require('react')
 var bulk = require('bulk-require')
 var _ = require('lodash')
-var commands = _.values(bulk(__dirname + '/', './*.js'))
+console.log(bulk(__dirname + '/../', '*/index.js'))
+var commands = _.values(bulk(__dirname + '/../', '*/index.js'))
 
 exports.name = "ls"
 exports.description = "Lists all available actions with descriptions"
 exports.execute = function() {
   return commands.map(function(command, i) {
+    if (!command.name || !command.description) return false
     return <div key={i}><b>{command.name}</b>: {command.description}</div>
   })
 }
